+++ src/main/java/duke/commands/AddTodoCommand.java
@@ -1,4 +1,9 @@
-import Tasks.Todo;
+package duke.commands;
+
+import duke.utils.Storage;
+import duke.utils.TaskList;
+import duke.tasks.Todo;
+import duke.utils.Ui;
 
 public class AddTodoCommand extends Command {
