+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.exceptions;
 
 public class DukeException extends Exception {
     public DukeException(String errorMessage) {
