+++ src/main/java/duke/exceptions/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.exceptions;
 
 public class EmptyDescriptionException extends DukeException {
