+++ src/main/java/duke/exceptions/EmptyTimestampException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.exceptions;
 
 public class EmptyTimestampException extends DukeException {
