+++ src/main/java/duke/exceptions/InvalidTaskNumberException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.exceptions;
 
 public class InvalidTaskNumberException extends DukeException {
