+++ src/main/java/duke/exceptions/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.exceptions;
 
 public class UnknownCommandException extends DukeException {
