+++ src/main/java/duke/tasks/Event.java
@@ -1,21 +1,17 @@
-package Tasks;
-import Exceptions.InvalidTimeStampException;
+package duke.tasks;
 
-import java.io.File;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
+import duke.exceptions.InvalidTimeStampException;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private static DateTimeFormatter inputFormatter = DateTimeFormatter.ofPattern("d/M/yyyy' 'HHmm");
     private static DateTimeFormatter outputFormatter = DateTimeFormatter.ofPattern("MMM d yyyy, h:mma");
 
     private LocalDateTime timeInfo;
-    public Deadline(String taskDetails, String timeInfoString) throws InvalidTimeStampException {
+    public Event(String taskDetails, String timeInfoString) throws InvalidTimeStampException {
         super(taskDetails);
         try {
             this.timeInfo = LocalDateTime.parse(timeInfoString, inputFormatter);
@@ -24,14 +20,15 @@ public class Deadline extends Task {
         }
     }
 
+    @Override
     public String toString() {
         String timeInfoString = timeInfo.format(outputFormatter);
-        return String.format("[D]%s (by: %s)", super.toString(), timeInfoString);
+        return String.format("[E]%s (at: %s)", super.toString(), timeInfoString);
     }
 
     @Override
     public String dataToString() {
         String timeInfoString = timeInfo.format(inputFormatter);
-        return String.format("D | %s | %s", super.dataToString(), timeInfoString);
+        return String.format("E | %s | %s", super.dataToString(), timeInfoString);
     }
 }
