+++ src/main/java/duke/tasks/Todo.java
@@ -1,6 +1,4 @@
-package Tasks;
-
-import java.io.File;
+package duke.tasks;
 
 public class Todo extends Task {
     public Todo(String taskDetails) {
